build: Rename x86_64 CMOS RTC driver to `cmos`
